feat: include multiple platforms in plugin.yaml
This change use goreleaser to build the arch and os binary matrix, and also to create
a changelog for the project and interact with GitHub (although this is not integrated
as of yet).

Shell scripts have been added to hack/ for editing convenience as inputs require
validation, and such implementation doesn't belong in a Makefile.