feat: add init command
This change adds the "helm dump init" command. This command currently only works in the case a
valid connection with a Kubernetes or OpenShift cluster is available, and will produce a valid chart.

Tests should be decoupled from the availability of a cluster using a fake dynamic client, so automated
tests can be self-contained.