feat: introduce helm_dump_init plugin
This plugin is intended to be used by crane-lib transform API in the resource extraction
pipeline. There's two cases being currently covered: when the input object contains labels
and when it doesn't contain labels. This is required as JSONPatch works incrementally and
won't auto-vivify the ".metadata.labels" map[string]interface{}{} on our behalf.