refactor: updateTemplate now uses the AST to perform modifications
This change caches the resource template whenever it first tries to modify it. This is required to keep a valid YAML representation of the original resource so modifications can happen to those instead of trying to re-use an invalid document.

updateTemplate uses a visitor to identify and collect the information necessary to change the document, particularly the beginOffset and endOffset in bytes that must be replaced in the YAML document.